refactor: CustomInquiryRepository 로 설정
