+++ src/main/java/com/zerobase/babdeusilbun/repository/InquiryRepository.java
@@ -1,7 +1,8 @@
 package com.zerobase.babdeusilbun.repository;
 
 import com.zerobase.babdeusilbun.domain.Inquiry;
+import com.zerobase.babdeusilbun.repository.custom.CustomInquiryRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InquiryRepository extends JpaRepository<Inquiry, Long> {
+public interface InquiryRepository extends JpaRepository<Inquiry, Long>, CustomInquiryRepository {
 }
